+++ src/lib/controls/dropdown-multiselect/dropdownMultiselectComponent.angular.spec.ts
@@ -316,16 +316,53 @@ describe("[DropdownMultiSelectComponent]", () => {
         });
 
         describe("when the list has been filtered", () => {
+            let showAll;
             beforeEach(() => {
+                jasmine.clock().uninstall();
+                jasmine.clock().install();
                 component.filterTextboxComponent.value = "2";
                 fixture.detectChanges();
+                jasmine.clock().tick(100);
+                showAll = rootElement.query(By.css('.dropdown-item--select-all'));
             });
-            it("the select all item text is 'Välj alla'", () => {
-                let item = rootElement.query(By.css('.dropdown-item--select-all'));
-                let itemLink = item.query(By.css("a"));
-                expect(itemLink.nativeElement.title).toBe(component.showAllItemText);
+            it('there are five focusable items', () => {
+                jasmine.clock().tick(200);
+                expect(component.focusableItems.length).toBe(5);
             });
+            describe('and space is pressed', () => {
+                beforeEach(() => {
+                    jasmine.clock().uninstall();
+                    jasmine.clock().install();
+                    showAll.triggerEventHandler('keydown', { keyCode: 32, preventDefault: function () { } } as KeyboardEvent);
+                    fixture.detectChanges();
+                });
 
+                it('filter is cleared', () => {
+                    jasmine.clock().tick(100);
+                    expect(component.filter).toBe('');
+                });
+                it('there are 23 focusable items', () => {
+                    jasmine.clock().tick(100);
+                    expect(component.focusableItems.length).toBe(23);
+                });
+            });
+            describe('and enter is pressed', () => {
+                beforeEach(() => {
+                    jasmine.clock().uninstall();
+                    jasmine.clock().install();
+                    showAll.triggerEventHandler('keydown', { keyCode: 13, preventDefault: function () { } } as KeyboardEvent);
+                    fixture.detectChanges();
+                });
+
+                it('filter is cleared', () => {
+                    jasmine.clock().tick(100);
+                    expect(component.filter).toBe('');
+                });
+                it('there are 23 focusable items', () => {
+                    jasmine.clock().tick(100);
+                    expect(component.focusableItems.length).toBe(23);
+                });
+            });
         });
 
         describe("when all items are selected and the list is filtered", () => {
@@ -507,7 +544,8 @@ describe("[DropdownMultiSelectComponent]", () => {
             });
         });
     });
-    describe('Compoennt is reseted', () => {
+
+    describe('Component is reseted', () => {
         beforeEach(() => {
             component.writeValue(null);
         });
@@ -518,5 +556,273 @@ describe("[DropdownMultiSelectComponent]", () => {
             });
         });
     });
-});
 
+    describe('WCAG Tests', () => {
+        let dropdownElement: DebugElement;
+        let listElement: DebugElement;
+        beforeEach(() => {
+            dropdownElement = rootElement.query(By.css('.dropdown--edit'));
+            component.ngOnChanges();
+            fixture.detectChanges();
+        });
+
+        it('The dropdown has role combobox.', () => {
+            expect(dropdownElement.attributes['role']).toBe('combobox');
+        });
+
+        it('Aria-expanded is false', () => {
+            expect(dropdownElement.attributes['aria-expanded']).toBe('false');
+        });
+        it('aria-disabled is false', () => {
+            expect(dropdownElement.attributes['aria-disabled']).toBe('false');
+        });
+        it('tabindex to be 0', () => {
+            expect(dropdownElement.properties['tabIndex']).toBe('0');
+        });
+
+        describe('list is initialized with items', () => {
+            let listItems;
+            beforeEach(() => {
+                component.expanded = false;
+
+                component.items = [{ displayName: 'one', value: 1 }, { displayName: 'two', value: 2 }, { displayName: 'three', value: 3 }];
+                fixture.detectChanges();
+
+                component.ngOnChanges();
+                listElement = rootElement.query(By.css('.dropdown__menu__items'));
+                dropdownElement = rootElement.query(By.css('.dropdown--edit'));
+                listItems = rootElement.queryAll(By.css('.dropdown-item'));
+            });
+            it('the list has role listbox', () => {
+                expect(listElement.attributes['role']).toBe('listbox');
+            });
+            it('has four focusable items', () => {
+                expect(component.focusableItems.length).toBe(4);
+            });
+            it('first item is "Välja alla"', () => {
+                expect(listElement.children[0].classes['dropdown-item--select-all']).toBe(true);
+            });
+            it('there are 5 focusable items ', () => {
+                expect(component.focusableItems.length).toBe(4);
+            })
+            describe('and alt + arrow down is pressed', () => {
+                beforeEach(() => {
+                    dropdownElement.triggerEventHandler('keydown', { keyCode: 40, altKey: true, preventDefault: function () { } } as KeyboardEvent);
+                    fixture.detectChanges();
+                });
+                it('dropdown is collapsed', () => {
+                    expect(dropdownElement.classes['dropdown--open']).toBe(true);
+                });
+                describe('and alt + arrow up is pressed', () => {
+                    beforeEach(() => {
+                        dropdownElement.triggerEventHandler('keydown', { keyCode: 38, altKey: true, preventDefault: function () { } } as KeyboardEvent);
+                        fixture.detectChanges();
+                    });
+                    it('dropdown is collapsed', () => {
+                        expect(dropdownElement.classes['dropdown--open']).toBe(false);
+                    });
+                });
+            });
+            describe('and space is pressed', () => {
+                beforeEach(() => {
+                    dropdownElement.triggerEventHandler('keydown', { keyCode: 32, preventDefault: function () { } } as KeyboardEvent);
+                    fixture.detectChanges();
+                });
+                it('dropdown is expanded', () => {
+                    expect(dropdownElement.classes['dropdown--open']).toBe(true);
+                });
+
+                describe('key arrow down focuses first item', () => {
+                    beforeEach(() => {
+                        spyOn(component.focusableItems[0], 'focus').and.callThrough();
+                        dropdownElement.triggerEventHandler('keydown', { keyCode: 40, preventDefault: function () { } } as KeyboardEvent);
+                        fixture.detectChanges();
+                    });
+
+                    it('first element has focus', () => {
+                        let focusedElement = rootElement.query(By.css(':focus'));
+                        expect(focusedElement.children[0].properties['title']).toBe('Välj alla');
+                    });
+                    it('first element has aria label "Välj alla"', () => {
+                        expect(listElement.children[0].attributes['aria-label']).toBe('Välj alla');
+                    });
+                    it('focusable items first object has focus', () => {
+                        expect(component.focusableItems[0].focus).toHaveBeenCalledTimes(1);
+                    });
+                    it('and select all items checkbox is marked', () => {
+                        expect(component.selectAllItemsMarked).toBe(true);
+                    });
+                    it('and select all items checkbox is not checked', () => {
+                        expect(component.selectAllItemsChecked).toBe(false);
+                    });
+                    describe('enter is pressed ', () => {
+                        let selectAll;
+                        beforeEach(() => {
+                            selectAll = rootElement.query(By.css('.dropdown-item--select-all'));
+                            selectAll.triggerEventHandler('keydown', { keyCode: 13, preventDefault: function () { } } as KeyboardEvent);
+                        });
+                        it('and select all items checkbox is checked', () => {
+                            expect(component.selectAllItemsChecked).toBe(true);
+                        });
+                        describe('enter is pressed again ', () => {
+                            beforeEach(() => {
+                                selectAll.triggerEventHandler('keydown', { keyCode: 13, preventDefault: function () { } } as KeyboardEvent);
+                            });
+                            it('and select all items checkbox is unchecked', () => {
+                                expect(component.selectAllItemsChecked).toBe(false);
+                            });
+                        });
+                    });
+                    describe('space is pressed ', () => {
+                        let selectAll;
+                        beforeEach(() => {
+                            selectAll = rootElement.query(By.css('.dropdown-item--select-all'));
+                            selectAll.triggerEventHandler('keydown', { keyCode: 32, preventDefault: function () { } } as KeyboardEvent);
+                        });
+                        it('and select all items checkbox is checked', () => {
+                            expect(component.selectAllItemsChecked).toBe(true);
+                        });
+                    });
+
+                    describe('key arrow down againn, focus on next item', () => {
+                        beforeEach(() => {
+                            spyOn(component.focusableItems[1], 'focus').and.callThrough();
+                            dropdownElement.triggerEventHandler('keydown', { keyCode: 40, preventDefault: function () { } } as KeyboardEvent);
+                            fixture.detectChanges();
+                        });
+                        it('second focusable element has aria label "one"', () => {
+                            expect(listElement.children[1].attributes['aria-label']).toBe('one');
+                        });
+                        it('focusable items first object has focus', () => {
+                            expect(component.focusableItems[1].focus).toHaveBeenCalledTimes(1);
+                        });
+                        describe('key arrow down pressed three times, focus is back to first element', () => {
+                            beforeEach(() => {
+                                dropdownElement.triggerEventHandler('keydown', { keyCode: 40, preventDefault: function () { } } as KeyboardEvent);
+                                dropdownElement.triggerEventHandler('keydown', { keyCode: 40, preventDefault: function () { } } as KeyboardEvent);
+                                dropdownElement.triggerEventHandler('keydown', { keyCode: 40, preventDefault: function () { } } as KeyboardEvent);
+                                fixture.detectChanges();
+                            });
+                            it('focusable items first object has focus', () => {
+                                expect(component.focusableItems[0].focus).toHaveBeenCalledTimes(2);
+                            });
+                            it('and select all items checkbox is marked', () => {
+                                expect(component.selectAllItemsMarked).toBe(true);
+                            });
+                            it('and select all items checkbox is not checked', () => {
+                                expect(component.selectAllItemsChecked).toBe(false);
+                            });
+                        });
+                        describe('and tab is pressed', () => {
+                            beforeEach(() => {
+                                listItems[1].triggerEventHandler('keydown', { keyCode: 9, preventDefault: function () { } } as KeyboardEvent);
+                                dropdownElement.triggerEventHandler('keydown', { keyCode: 9, preventDefault: function () { } } as KeyboardEvent);
+                                fixture.detectChanges();
+                            });
+
+                            it('current element is not selected', () => {
+                                expect(component.items[1].selected).toBeFalsy();
+                            });
+
+                            it('dropdown is collapsed', () => {
+                                expect(dropdownElement.classes['dropdown--open']).toBe(false);
+                            });
+                        });
+                    });
+                });
+                describe('key arrow up focuses last item', () => {
+                    beforeEach(() => {
+                        spyOn(component.focusableItems[3], 'focus').and.callThrough();
+                        dropdownElement.triggerEventHandler('keydown', { keyCode: 38, preventDefault: function () { } } as KeyboardEvent);
+                        listItems[2].triggerEventHandler('focusin', {});
+                        fixture.detectChanges();
+                    });
+                    it('last element has aria label "Välj alla"', () => {
+                        expect(listElement.children[3].attributes['aria-label']).toBe('three');
+                    });
+                    it('focusable items last object has focus', () => {
+                        expect(component.focusableItems[3].focus).toHaveBeenCalledTimes(1);
+                    });
+                    it('and last item checkbox is marked', () => {
+                        expect(component.items[2].marked).toBe(true);
+                    });
+                    it('and last item is not selected', () => {
+                        expect(component.items[2].selected).toBeFalsy();
+                    });
+
+                    describe('enter is pressed', () => {
+                        beforeEach(() => {
+                            listItems[2].triggerEventHandler('keydown', { keyCode: 13, preventDefault: function () { } } as KeyboardEvent);
+                            fixture.detectChanges();
+                        });
+                        it('and last item checkbox is selected', () => {
+                            expect(component.items[2].selected).toBe(true);
+                        });
+                    });
+                    describe('space is pressed', () => {
+                        beforeEach(() => {
+                            listItems[2].triggerEventHandler('keydown', { keyCode: 32, preventDefault: function () { } } as KeyboardEvent);
+                            fixture.detectChanges();
+                        });
+                        it('and last item checkbox is unselected', () => {
+                            expect(component.items[2].selected).toBe(true);
+                        });
+
+                    });
+
+                    describe('key arrow up again, focus on second item', () => {
+                        beforeEach(() => {
+                            spyOn(component.focusableItems[2], 'focus').and.callThrough();
+                            dropdownElement.triggerEventHandler('keydown', { keyCode: 38, preventDefault: function () { } } as KeyboardEvent);
+                            listItems[1].triggerEventHandler('focusin', {});
+                            fixture.detectChanges();
+                        });
+                        it('focusable items first object has focus', () => {
+                            expect(component.focusableItems[2].focus).toHaveBeenCalledTimes(1);
+                        });
+                        it('and second last item checkbox is marked', () => {
+                            expect(component.items[1].marked).toBe(true);
+                        });
+                    });
+                });
+            });
+        });
+
+        describe('dropdown initialized as readonly', () => {
+            let dropdownElement: DebugElement;
+            let readonlyDropdownElement: DebugElement;
+            let selectedItemsElement: DebugElement[];
+            beforeEach(() => {
+                component.expanded = false;
+
+                component.items = [{ displayName: 'one', value: 1, selected: true }, { displayName: 'two', value: 2 }, { displayName: 'three', value: 3 }];
+                component.readonly = true;
+                dropdownElement = rootElement.query(By.css('.dropdown--edit'));
+                readonlyDropdownElement = rootElement.query(By.css('.dropdown__multiselect-readonlylist'));
+
+                component.ngOnChanges();
+                fixture.detectChanges();
+                selectedItemsElement = readonlyDropdownElement.queryAll(By.css('li'));
+            });
+
+            it('aria-readonly set to true on readonly-items', () => {
+                expect(readonlyDropdownElement.attributes['aria-readonly']).toBe('true');
+            })
+
+            it('tabindex to be 0 on readonly-list', () => {
+                expect(readonlyDropdownElement.properties['tabIndex']).toBe('0');
+            });
+            it('tabindex to be -1 on dropdown', () => {
+                expect(dropdownElement.properties['tabIndex']).toBe('-1');
+            });
+
+            it('selectedItem is "one"', () => {
+                expect(selectedItemsElement.length).toBe(1);
+            })
+
+            it('selectedItem is "one"', () => {
+                expect(selectedItemsElement[0].nativeElement.innerText).toBe('one');
+            })
+        });
+    })
+});